+++ src/routes/auth/reset/[token].svelte
@@ -1,0 +1,3 @@
+<script lang="ts">
+  // TODO: Fields: password, password confirmation and button
+</script>