+++ src/routes/info.svelte
@@ -2,5 +2,9 @@
   export const prerender = true
 </script>
 
+<svelte:head>
+  <title>Info Page</title>
+</svelte:head>
+
 <h1>Info</h1>
 <h4>Public</h4>